chore: move map-content to components
